requirements.txt
click==7.1.2
cv==1.0.0
Flask==1.1.2
Flask-Cors==3.0.10
imageio==2.9.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.0
opencv-python==4.5.1.48
Pillow==8.1.0
six==1.15.0
Werkzeug==1.0.1
matplotlib~=3.3.4
scipy~=1.6.0
scikit-image~=0.18.1